feat(avatar): update demo
