+++ src/avatar/_example/style/index.less
@@ -1,50 +1,3 @@
-// .tdesign-mobile-demo {
-//   background-color: #f6f6f6;
-
-//   .title {
-//     padding: 24px 16px 8px;
-//     opacity: 1;
-//     color: #000000e6;
-//     font-size: 20px;
-//     font-weight: 700;
-//     line-height: 28px;
-//   }
-
-//   .summary {
-//     opacity: 1;
-//     color: #0006;
-//     font-size: 12px;
-//     line-height: 22px;
-//     padding: 0 16px;
-//   }
-
-//   &-block {
-//     width: 100%;
-
-//     &__title {
-//       padding: 24px 16px 0;
-//       font-size: 16px;
-//       font-weight: 700;
-//     }
-
-//     &__summary {
-//       opacity: 1;
-//       color: #0006;
-//       font-size: 12px;
-//       line-height: 22px;
-//       padding: 16px;
-//     }
-//     &__title + &__summary {
-//       padding: 8px 16px 16px;
-//     }
-//   }
-
-//   .avatar-demo {
-//     margin-left: 16px;
-//     margin-bottom: 16px;
-//   }
-// }
-
 .avatar-demo {
   margin-left: 16px;
   margin-bottom: 16px;
